Extract ContainerResult class
